% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{stat_windmill}
\alias{stat_windmill}
\title{A Cat Function}
\usage{
stat_windmill(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  span_x = 1,
  ...
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
